+++ daemons/controld/Makefile.am
@@ -1,5 +1,7 @@
 #
-# Copyright 2004-2018 Andrew Beekhof <andrew@beekhof.net>
+# Copyright 2018-2019 the Pacemaker project contributors
+#
+# The version control history for this file may have further details.
 #
 # This source code is licensed under the GNU General Public License version 2
 # or later (GPLv2+) WITHOUT ANY WARRANTY.
@@ -46,6 +48,7 @@ pacemaker_controld_SOURCES = pacemaker-controld.c	\
 			     controld_election.c	\
 			     controld_execd.c		\
 			     controld_execd_state.c	\
+			     controld_fencing.c		\
 			     controld_fsa.c		\
 			     controld_join_client.c	\
 			     controld_join_dc.c		\
