+++ daemons/controld/controld_fsa.c
@@ -26,6 +26,7 @@
 #include <pacemaker-controld.h>
 #include <controld_messages.h>
 #include <controld_fsa.h>
+#include <controld_fencing.h>
 #include <controld_transition.h>
 #include <controld_matrix.h>
