+++ daemons/controld/controld_messages.c
@@ -25,6 +25,7 @@
 #include <pacemaker-controld.h>
 #include <controld_messages.h>
 #include <controld_lrm.h>
+#include <controld_fencing.h>
 #include <controld_transition.h>
 #include <controld_throttle.h>
