+++ daemons/controld/controld_te_callbacks.c
@@ -17,6 +17,7 @@
 
 #include <controld_transition.h>
 #include <controld_fsa.h>
+#include <controld_fencing.h>
 
 #include <crm/cluster.h>        /* For ONLINESTATUS etc */
 
@@ -27,21 +28,9 @@ gboolean shuttingdown = FALSE;
 crm_graph_t *transition_graph;
 crm_trigger_t *transition_trigger = NULL;
 
-static unsigned long int stonith_max_attempts = 10;
-
 /* #define RSC_OP_TEMPLATE "//"XML_TAG_DIFF_ADDED"//"XML_TAG_CIB"//"XML_CIB_TAG_STATE"[@uname='%s']"//"XML_LRM_TAG_RSC_OP"[@id='%s]" */
 #define RSC_OP_TEMPLATE "//"XML_TAG_DIFF_ADDED"//"XML_TAG_CIB"//"XML_LRM_TAG_RSC_OP"[@id='%s']"
 
-void
-update_stonith_max_attempts(const char* value)
-{
-    if (safe_str_eq(value, CRM_INFINITY_S)) {
-       stonith_max_attempts = CRM_SCORE_INFINITY;
-    }
-    else {
-       stonith_max_attempts = crm_int_helper(value, NULL);
-    }
-}
 static void
 te_update_diff_v1(const char *event, xmlNode *diff)
 {
@@ -644,236 +633,6 @@ process_te_message(xmlNode * msg, xmlNode * xml_data)
     crm_trace("finished processing message");
 
     return TRUE;
-}
-
-GHashTable *stonith_failures = NULL;
-struct st_fail_rec {
-    int count;
-};
-
-static gboolean
-too_many_st_failures(const char *target)
-{
-    GHashTableIter iter;
-    const char *key = NULL;
-    struct st_fail_rec *value = NULL;
-
-    if (stonith_failures == NULL) {
-        return FALSE;
-    }
-
-    if (target == NULL) {
-        g_hash_table_iter_init(&iter, stonith_failures);
-        while (g_hash_table_iter_next(&iter, (gpointer *) & key, (gpointer *) & value)) {
-            if (value->count >= stonith_max_attempts) {
-                target = (const char*)key;
-                goto too_many;
-            }
-        }
-    } else {
-        value = g_hash_table_lookup(stonith_failures, target);
-        if ((value != NULL) && (value->count >= stonith_max_attempts)) {
-            goto too_many;
-        }
-    }
-    return FALSE;
-
-too_many:
-    crm_warn("Too many failures (%d) to fence %s, giving up",
-             value->count, target);
-    return TRUE;
-}
-
-/*!
- * \internal
- * \brief Reset a stonith fail count
- *
- * \param[in] target  Name of node to reset, or NULL for all
- */
-void
-st_fail_count_reset(const char *target)
-{
-    if (stonith_failures == NULL) {
-        return;
-    }
-
-    if (target) {
-        struct st_fail_rec *rec = NULL;
-
-        rec = g_hash_table_lookup(stonith_failures, target);
-        if (rec) {
-            rec->count = 0;
-        }
-    } else {
-        GHashTableIter iter;
-        const char *key = NULL;
-        struct st_fail_rec *rec = NULL;
-
-        g_hash_table_iter_init(&iter, stonith_failures);
-        while (g_hash_table_iter_next(&iter, (gpointer *) &key,
-                                      (gpointer *) &rec)) {
-            rec->count = 0;
-        }
-    }
-}
-
-void
-st_fail_count_increment(const char *target)
-{
-    struct st_fail_rec *rec = NULL;
-
-    if (stonith_failures == NULL) {
-        stonith_failures = crm_str_table_new();
-    }
-
-    rec = g_hash_table_lookup(stonith_failures, target);
-    if (rec) {
-        rec->count++;
-    } else {
-        rec = malloc(sizeof(struct st_fail_rec));
-        if(rec == NULL) {
-            return;
-        }
-
-        rec->count = 1;
-        g_hash_table_insert(stonith_failures, strdup(target), rec);
-    }
-}
-
-/*!
- * \internal
- * \brief Abort transition due to stonith failure
- *
- * \param[in] abort_action  Whether to restart or stop transition
- * \param[in] target  Don't restart if this (NULL for any) has too many failures
- * \param[in] reason  Log this stonith action XML as abort reason (or NULL)
- */
-void
-abort_for_stonith_failure(enum transition_action abort_action,
-                          const char *target, xmlNode *reason)
-{
-    /* If stonith repeatedly fails, we eventually give up on starting a new
-     * transition for that reason.
-     */
-    if ((abort_action != tg_stop) && too_many_st_failures(target)) {
-        abort_action = tg_stop;
-    }
-    abort_transition(INFINITY, abort_action, "Stonith failed", reason);
-}
-
-void
-tengine_stonith_callback(stonith_t * stonith, stonith_callback_data_t * data)
-{
-    char *uuid = NULL;
-    int stonith_id = -1;
-    int transition_id = -1;
-    crm_action_t *action = NULL;
-    int call_id = data->call_id;
-    int rc = data->rc;
-    char *userdata = data->userdata;
-
-    CRM_CHECK(userdata != NULL, return);
-    crm_notice("Stonith operation %d/%s: %s (%d)", call_id, (char *)userdata,
-               pcmk_strerror(rc), rc);
-
-    if (AM_I_DC == FALSE) {
-        return;
-    }
-
-    /* crm_info("call=%d, optype=%d, node_name=%s, result=%d, node_list=%s, action=%s", */
-    /*       op->call_id, op->optype, op->node_name, op->op_result, */
-    /*       (char *)op->node_list, op->private_data); */
-
-    /* filter out old STONITH actions */
-    CRM_CHECK(decode_transition_key(userdata, &uuid, &transition_id, &stonith_id, NULL),
-              goto bail);
-
-    if (transition_graph->complete || stonith_id < 0 || safe_str_neq(uuid, te_uuid)
-        || transition_graph->id != transition_id) {
-        crm_info("Ignoring STONITH action initiated outside of the current transition");
-        goto bail;
-    }
-
-    action = get_action(stonith_id, FALSE);
-    if (action == NULL) {
-        crm_err("Stonith action not matched");
-        goto bail;
-    }
-
-    stop_te_timer(action->timer);
-    if (rc == pcmk_ok) {
-        const char *target = crm_element_value(action->xml, XML_LRM_ATTR_TARGET);
-        const char *uuid = crm_element_value(action->xml, XML_LRM_ATTR_TARGET_UUID);
-        const char *op = crm_meta_value(action->params, "stonith_action"); 
-
-        crm_info("Stonith operation %d for %s passed", call_id, target);
-        if (action->confirmed == FALSE) {
-            te_action_confirmed(action);
-            if (safe_str_eq("on", op)) {
-                const char *value = NULL;
-                char *now = crm_itoa(time(NULL));
-
-                update_attrd(target, CRM_ATTR_UNFENCED, now, NULL, FALSE);
-                free(now);
-
-                value = crm_meta_value(action->params, XML_OP_ATTR_DIGESTS_ALL);
-                update_attrd(target, CRM_ATTR_DIGESTS_ALL, value, NULL, FALSE);
-
-                value = crm_meta_value(action->params, XML_OP_ATTR_DIGESTS_SECURE);
-                update_attrd(target, CRM_ATTR_DIGESTS_SECURE, value, NULL, FALSE);
-
-            } else if (action->sent_update == FALSE) {
-                send_stonith_update(action, target, uuid);
-                action->sent_update = TRUE;
-            }
-        }
-        st_fail_count_reset(target);
-
-    } else {
-        const char *target = crm_element_value(action->xml, XML_LRM_ATTR_TARGET);
-        enum transition_action abort_action = tg_restart;
-
-        action->failed = TRUE;
-        crm_notice("Stonith operation %d for %s failed (%s): aborting transition.",
-                   call_id, target, pcmk_strerror(rc));
-
-        /* If no fence devices were available, there's no use in immediately
-         * checking again, so don't start a new transition in that case.
-         */
-        if (rc == -ENODEV) {
-            crm_warn("No devices found in cluster to fence %s, giving up",
-                     target);
-            abort_action = tg_stop;
-        }
-
-        /* Increment the fail count now, so abort_for_stonith_failure() can
-         * check it. Non-DC nodes will increment it in tengine_stonith_notify().
-         */
-        st_fail_count_increment(target);
-        abort_for_stonith_failure(abort_action, target, NULL);
-    }
-
-    update_graph(transition_graph, action);
-    trigger_graph();
-
-  bail:
-    free(userdata);
-    free(uuid);
-    return;
-}
-
-void
-cib_fencing_updated(xmlNode * msg, int call_id, int rc, xmlNode * output, void *user_data)
-{
-    if (rc < pcmk_ok) {
-        crm_err("Fencing update %d for %s: failed - %s (%d)",
-                call_id, (char *)user_data, pcmk_strerror(rc), rc);
-        crm_log_xml_warn(msg, "Failed update");
-        abort_transition(INFINITY, tg_shutdown, "CIB update failed", NULL);
-
-    } else {
-        crm_info("Fencing update %d for %s: complete", call_id, (char *)user_data);
-    }
 }
 
 void
