+++ daemons/controld/controld_te_utils.c
@@ -6,441 +6,14 @@
  */
 
 #include <crm_internal.h>
-
-#include <sys/param.h>
 #include <crm/crm.h>
-
 #include <crm/msg_xml.h>
-
 #include <crm/common/xml.h>
+
 #include <controld_transition.h>
 #include <controld_fsa.h>
-#include <controld_lrm.h>
 #include <controld_messages.h>
 #include <controld_throttle.h>
-#include <crm/fencing/internal.h>
-
-crm_trigger_t *stonith_reconnect = NULL;
-static crm_trigger_t *stonith_history_sync_trigger = NULL;
-static mainloop_timer_t *stonith_history_sync_timer = NULL;
-
-/*
- * stonith cleanup list
- *
- * If the DC is shot, proper notifications might not go out.
- * The stonith cleanup list allows the cluster to (re-)send
- * notifications once a new DC is elected.
- */
-
-static GListPtr stonith_cleanup_list = NULL;
-
-/*!
- * \internal
- * \brief Add a node to the stonith cleanup list
- *
- * \param[in] target  Name of node to add
- */
-void
-add_stonith_cleanup(const char *target) {
-    stonith_cleanup_list = g_list_append(stonith_cleanup_list, strdup(target));
-}
-
-/*!
- * \internal
- * \brief Remove a node from the stonith cleanup list
- *
- * \param[in] Name of node to remove
- */
-void
-remove_stonith_cleanup(const char *target)
-{
-    GListPtr iter = stonith_cleanup_list;
-
-    while (iter != NULL) {
-        GListPtr tmp = iter;
-        char *iter_name = tmp->data;
-
-        iter = iter->next;
-        if (safe_str_eq(target, iter_name)) {
-            crm_trace("Removing %s from the cleanup list", iter_name);
-            stonith_cleanup_list = g_list_delete_link(stonith_cleanup_list, tmp);
-            free(iter_name);
-        }
-    }
-}
-
-/*!
- * \internal
- * \brief Purge all entries from the stonith cleanup list
- */
-void
-purge_stonith_cleanup()
-{
-    if (stonith_cleanup_list) {
-        GListPtr iter = NULL;
-
-        for (iter = stonith_cleanup_list; iter != NULL; iter = iter->next) {
-            char *target = iter->data;
-
-            crm_info("Purging %s from stonith cleanup list", target);
-            free(target);
-        }
-        g_list_free(stonith_cleanup_list);
-        stonith_cleanup_list = NULL;
-    }
-}
-
-/*!
- * \internal
- * \brief Send stonith updates for all entries in cleanup list, then purge it
- */
-void
-execute_stonith_cleanup()
-{
-    GListPtr iter;
-
-    for (iter = stonith_cleanup_list; iter != NULL; iter = iter->next) {
-        char *target = iter->data;
-        crm_node_t *target_node = crm_get_peer(0, target);
-        const char *uuid = crm_peer_uuid(target_node);
-
-        crm_notice("Marking %s, target of a previous stonith action, as clean", target);
-        send_stonith_update(NULL, target, uuid);
-        free(target);
-    }
-    g_list_free(stonith_cleanup_list);
-    stonith_cleanup_list = NULL;
-}
-
-/* end stonith cleanup list functions */
-
-static gboolean
-fail_incompletable_stonith(crm_graph_t * graph)
-{
-    GListPtr lpc = NULL;
-    const char *task = NULL;
-    xmlNode *last_action = NULL;
-
-    if (graph == NULL) {
-        return FALSE;
-    }
-
-    for (lpc = graph->synapses; lpc != NULL; lpc = lpc->next) {
-        GListPtr lpc2 = NULL;
-        synapse_t *synapse = (synapse_t *) lpc->data;
-
-        if (synapse->confirmed) {
-            continue;
-        }
-
-        for (lpc2 = synapse->actions; lpc2 != NULL; lpc2 = lpc2->next) {
-            crm_action_t *action = (crm_action_t *) lpc2->data;
-
-            if (action->type != action_type_crm || action->confirmed) {
-                continue;
-            }
-
-            task = crm_element_value(action->xml, XML_LRM_ATTR_TASK);
-            if (task && safe_str_eq(task, CRM_OP_FENCE)) {
-                action->failed = TRUE;
-                last_action = action->xml;
-                update_graph(graph, action);
-                crm_notice("Failing action %d (%s): fencer terminated",
-                           action->id, ID(action->xml));
-            }
-        }
-    }
-
-    if (last_action != NULL) {
-        crm_warn("Fencer failure resulted in unrunnable actions");
-        abort_for_stonith_failure(tg_restart, NULL, last_action);
-        return TRUE;
-    }
-
-    return FALSE;
-}
-
-static void
-tengine_stonith_connection_destroy(stonith_t * st, stonith_event_t * e)
-{
-    if (is_set(fsa_input_register, R_ST_REQUIRED)) {
-        crm_crit("Fencing daemon connection failed");
-        mainloop_set_trigger(stonith_reconnect);
-
-    } else {
-        crm_info("Fencing daemon disconnected");
-    }
-
-    /* cbchan will be garbage at this point, arrange for it to be reset */
-    if(stonith_api) {
-        stonith_api->state = stonith_disconnected;
-    }
-
-    if (AM_I_DC) {
-        fail_incompletable_stonith(transition_graph);
-        trigger_graph();
-    }
-}
-
-char *te_client_id = NULL;
-
-#ifdef HAVE_SYS_REBOOT_H
-#  include <unistd.h>
-#  include <sys/reboot.h>
-#endif
-
-static void
-tengine_stonith_notify(stonith_t * st, stonith_event_t * st_event)
-{
-    if(te_client_id == NULL) {
-        te_client_id = crm_strdup_printf("%s.%lu", crm_system_name,
-                                         (unsigned long) getpid());
-    }
-
-    if (st_event == NULL) {
-        crm_err("Notify data not found");
-        return;
-    }
-
-    crmd_alert_fencing_op(st_event);
-
-    if (st_event->result == pcmk_ok && safe_str_eq("on", st_event->action)) {
-        crm_notice("%s was successfully unfenced by %s (at the request of %s)",
-                   st_event->target, st_event->executioner ? st_event->executioner : "<anyone>", st_event->origin);
-                /* TODO: Hook up st_event->device */
-        return;
-
-    } else if (safe_str_eq("on", st_event->action)) {
-        crm_err("Unfencing of %s by %s failed: %s (%d)",
-                st_event->target, st_event->executioner ? st_event->executioner : "<anyone>",
-                pcmk_strerror(st_event->result), st_event->result);
-        return;
-
-    } else if (st_event->result == pcmk_ok && crm_str_eq(st_event->target, fsa_our_uname, TRUE)) {
-        crm_crit("We were allegedly just fenced by %s for %s!",
-                 st_event->executioner ? st_event->executioner : "<anyone>", st_event->origin); /* Dumps blackbox if enabled */
-
-        qb_log_fini(); /* Try to get the above log message to disk - somehow */
-
-        /* Get out ASAP and do not come back up.
-         *
-         * Triggering a reboot is also not the worst idea either since
-         * the rest of the cluster thinks we're safely down
-         */
-
-#ifdef RB_HALT_SYSTEM
-        reboot(RB_HALT_SYSTEM);
-#endif
-
-        /*
-         * If reboot() fails or is not supported, coming back up will
-         * probably lead to a situation where the other nodes set our
-         * status to 'lost' because of the fencing callback and will
-         * discard subsequent election votes with:
-         *
-         * Election 87 (current: 5171, owner: 103): Processed vote from east-03 (Peer is not part of our cluster)
-         *
-         * So just stay dead, something is seriously messed up anyway.
-         *
-         */
-        exit(CRM_EX_FATAL); // None of our wrappers since we already called qb_log_fini()
-        return;
-    }
-
-    /* Update the count of stonith failures for this target, in case we become
-     * DC later. The current DC has already updated its fail count in
-     * tengine_stonith_callback().
-     */
-    if (!AM_I_DC && safe_str_eq(st_event->operation, T_STONITH_NOTIFY_FENCE)) {
-        if (st_event->result == pcmk_ok) {
-            st_fail_count_reset(st_event->target);
-        } else {
-            st_fail_count_increment(st_event->target);
-        }
-    }
-
-    crm_notice("Peer %s was%s terminated (%s) by %s on behalf of %s: %s "
-               CRM_XS " initiator=%s ref=%s",
-               st_event->target, st_event->result == pcmk_ok ? "" : " not",
-               st_event->action,
-               st_event->executioner ? st_event->executioner : "<anyone>",
-               (st_event->client_origin? st_event->client_origin : "<unknown>"),
-               pcmk_strerror(st_event->result),
-               st_event->origin, st_event->id);
-
-    if (st_event->result == pcmk_ok) {
-        crm_node_t *peer = crm_find_known_peer_full(0, st_event->target, CRM_GET_PEER_ANY);
-        const char *uuid = NULL;
-        gboolean we_are_executioner = safe_str_eq(st_event->executioner, fsa_our_uname);
-
-        if (peer == NULL) {
-            return;
-        }
-
-        uuid = crm_peer_uuid(peer);
-
-        crm_trace("target=%s dc=%s", st_event->target, fsa_our_dc);
-        if(AM_I_DC) {
-            /* The DC always sends updates */
-            send_stonith_update(NULL, st_event->target, uuid);
-
-            /* @TODO Ideally, at this point, we'd check whether the fenced node
-             * hosted any guest nodes, and call remote_node_down() for them.
-             * Unfortunately, the controller doesn't have a simple, reliable way
-             * to map hosts to guests. It might be possible to track this in the
-             * peer cache via crm_remote_peer_cache_refresh(). For now, we rely
-             * on the PE creating fence pseudo-events for the guests.
-             */
-
-            if (st_event->client_origin && safe_str_neq(st_event->client_origin, te_client_id)) {
-
-                /* Abort the current transition graph if it wasn't us
-                 * that invoked stonith to fence someone
-                 */
-                crm_info("External fencing operation from %s fenced %s", st_event->client_origin, st_event->target);
-                abort_transition(INFINITY, tg_restart, "External Fencing Operation", NULL);
-            }
-
-            /* Assume it was our leader if we don't currently have one */
-        } else if (((fsa_our_dc == NULL) || safe_str_eq(fsa_our_dc, st_event->target))
-            && !is_set(peer->flags, crm_remote_node)) {
-
-            crm_notice("Target %s our leader %s (recorded: %s)",
-                       fsa_our_dc ? "was" : "may have been", st_event->target,
-                       fsa_our_dc ? fsa_our_dc : "<unset>");
-
-            /* Given the CIB resyncing that occurs around elections,
-             * have one node update the CIB now and, if the new DC is different,
-             * have them do so too after the election
-             */
-            if (we_are_executioner) {
-                send_stonith_update(NULL, st_event->target, uuid);
-            }
-            add_stonith_cleanup(st_event->target);
-        }
-
-        /* If the target is a remote node, and we host its connection,
-         * immediately fail all monitors so it can be recovered quickly.
-         * The connection won't necessarily drop when a remote node is fenced,
-         * so the failure might not otherwise be detected until the next poke.
-         */
-        if (is_set(peer->flags, crm_remote_node)) {
-            remote_ra_fail(st_event->target);
-        }
-
-        crmd_peer_down(peer, TRUE);
-     }
-}
-
-static gboolean
-do_stonith_history_sync(gpointer user_data)
-{
-    if (stonith_api && (stonith_api->state != stonith_disconnected)) {
-        stonith_history_t *history = NULL;
-
-        stonith_api->cmds->history(stonith_api,
-                                   st_opt_sync_call | st_opt_broadcast,
-                                   NULL, &history, 5);
-        stonith_history_free(history);
-        return TRUE;
-    } else {
-        crm_info("Skip triggering stonith history-sync as stonith is disconnected");
-        return FALSE;
-    }
-}
-
-static gboolean
-stonith_history_sync_set_trigger(gpointer user_data)
-{
-    mainloop_set_trigger(stonith_history_sync_trigger);
-    return FALSE;
-}
-
-void
-te_trigger_stonith_history_sync(void)
-{
-    /* trigger a sync in 5s to give more nodes the
-     * chance to show up so that we don't create
-     * unnecessary stonith-history-sync traffic
-     */
-
-    /* as we are finally checking the stonith-connection
-     * in do_stonith_history_sync we should be fine
-     * leaving stonith_history_sync_time & stonith_history_sync_trigger
-     * around
-     */
-    if (stonith_history_sync_trigger == NULL) {
-        stonith_history_sync_trigger =
-            mainloop_add_trigger(G_PRIORITY_LOW,
-                                 do_stonith_history_sync, NULL);
-    }
-
-    if(stonith_history_sync_timer == NULL) {
-        stonith_history_sync_timer =
-            mainloop_timer_add("history_sync", 5000,
-                               FALSE, stonith_history_sync_set_trigger,
-                               NULL);
-    }
-    crm_info("Fence history will be synchronized cluster-wide within 5 seconds");
-    mainloop_timer_start(stonith_history_sync_timer);
-}
-
-/*!
- * \brief Connect to fencer
- *
- * \param[in] user_data  If NULL, retry failures now, otherwise retry in main loop
- *
- * \return TRUE
- * \note If user_data is NULL, this will wait 2s between attempts, for up to
- *       30 attempts, meaning the controller could be blocked as long as 58s.
- */
-gboolean
-te_connect_stonith(gpointer user_data)
-{
-    int rc = pcmk_ok;
-
-    if (stonith_api == NULL) {
-        stonith_api = stonith_api_new();
-    }
-
-    if (stonith_api->state != stonith_disconnected) {
-        crm_trace("Already connected to fencer, no need to retry");
-        return TRUE;
-    }
-
-    if (user_data == NULL) {
-        // Blocking (retry failures now until successful)
-        rc = stonith_api_connect_retry(stonith_api, crm_system_name, 30);
-        if (rc != pcmk_ok) {
-            crm_err("Could not connect to fencer in 30 attempts: %s "
-                    CRM_XS " rc=%d", pcmk_strerror(rc), rc);
-        }
-    } else {
-        // Non-blocking (retry failures later in main loop)
-        rc = stonith_api->cmds->connect(stonith_api, crm_system_name, NULL);
-        if (rc != pcmk_ok) {
-            if (is_set(fsa_input_register, R_ST_REQUIRED)) {
-                crm_err("Fencer connection failed (will retry): %s "
-                        CRM_XS " rc=%d", pcmk_strerror(rc), rc);
-                mainloop_set_trigger(stonith_reconnect);
-            } else {
-                crm_info("Fencer connection failed (ignoring because no longer required): %s "
-                         CRM_XS " rc=%d", pcmk_strerror(rc), rc);
-            }
-            return TRUE;
-        }
-    }
-
-    if (rc == pcmk_ok) {
-        stonith_api->cmds->register_notification(stonith_api,
-                                                 T_STONITH_NOTIFY_DISCONNECT,
-                                                 tengine_stonith_connection_destroy);
-        stonith_api->cmds->register_notification(stonith_api,
-                                                 T_STONITH_NOTIFY_FENCE,
-                                                 tengine_stonith_notify);
-    }
-    return TRUE;
-}
 
 gboolean
 stop_te_timer(crm_action_timer_t * timer)
