+++ daemons/controld/controld_utils.h
@@ -85,10 +85,6 @@ int crmd_join_phase_count(enum crm_join_phase phase);
 void crmd_join_phase_log(int level);
 
 const char *get_timer_desc(fsa_timer_t * timer);
-void st_fail_count_reset(const char * target);
-void st_fail_count_increment(const char *target);
-void abort_for_stonith_failure(enum transition_action abort_action,
-                               const char *target, xmlNode *reason);
 void crmd_peer_down(crm_node_t *peer, bool full);
 unsigned int cib_op_timeout(void);
